# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API Documentation
  version: 1.0.0
  description: >
    REST API for the TaskFlow student project. All successful API responses use
    `{ statusCode, success, message, data }`; errors use
    `{ statusCode, success, message, errors }`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user account
      description: >
        **Purpose**: Registers a new user account in the TaskFlow system.<br>
        **Authorization**: Public (No authentication required).<br>
        **Side Effects**: Creates a new user record in MongoDB with hashed password, generates a temporary SHA-256 email verification token, and sends a verification email.<br>
        **Expected Behavior**: Returns HTTP 201 with the created user object omitting sensitive fields. Throws HTTP 409 if username or email already exists.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        "201":
          description: User account created successfully. Verification email dispatched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                success: true
                message: User registered successfully. Please check your email to verify your account.
                data:
                  user:
                    _id: 650c1f1e2f8b9a001c2d3e4f
                    username: alex_rivera
                    email: alex.rivera@taskflow.dev
                    isEmailVerified: false
        "400":
          $ref: '#/components/responses/ValidationError'
        "409":
          $ref: '#/components/responses/Conflict'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user credentials
      description: >
        **Purpose**: Authenticates a user using email or username along with their password.<br>
        **Authorization**: Public (No authentication required).<br>
        **Side Effects**: Saves the refresh token in MongoDB and sets HTTP-only `accessToken` and `refreshToken` cookies. In local development they are sent without the `Secure` flag; production uses secure, strict cookies.<br>
        **Expected Behavior**: Returns HTTP 200 with a user profile and access token. The refresh token is never returned in JSON.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
      responses:
        "200":
          description: User authenticated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: User logged in successfully
                data:
                  user:
                    _id: 650c1f1e2f8b9a001c2d3e4f
                    username: alex_rivera
                    email: alex.rivera@taskflow.dev
                    fullName: Alex Rivera
                    isEmailVerified: true
                    avatar:
                      url: https://placehold.co/200?text=No+Image
                      localPath: ""
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NTBjMWYxZTJmOGI5YTAwMWMyZDNlNGYiLCJ1c2VybmFtZSI6ImFsZXhfcml2ZXJhIiwiZW1haWwiOiJhbGV4LnJpdmVyYUB0YXNrZmxvdy5kZXYiLCJpYXQiOjE3Njg0NjYwMDAsImV4cCI6MTc2ODU1MjQwMH0.signature
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /auth/verify-email/{verificationToken}:
    get:
      tags:
        - Authentication
      summary: Verify user email address
      description: >
        **Purpose**: Validates the email verification token sent via email and marks user account as verified.<br>
        **Authorization**: Public.<br>
        **Side Effects**: Sets `isEmailVerified: true` and clears `emailVerificationToken` and `emailVerificationTokenExpiry` on User document.<br>
        **Expected Behavior**: Returns HTTP 200 on successful verification. Throws HTTP 400 if token is invalid or expired.
      operationId: verifyEmail
      parameters:
        - name: verificationToken
          in: path
          required: true
          description: Opaque temporary token received in the frontend verification link
          schema:
            type: string
          example: 4a2b3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b
      responses:
        "200":
          description: User email verified successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Email verified successfully
                data:
                  user:
                    isEmailVerified: true
        "400":
          $ref: '#/components/responses/ValidationError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access and refresh tokens
      description: >
        **Purpose**: Issues a fresh JWT Access Token and rotates the Refresh Token using an active Refresh Token.<br>
        **Authorization**: Public (Accepts Refresh Token via cookie or request body).<br>
        **Side Effects**: Updates `refreshToken` field in MongoDB user document and resets cookie.<br>
        **Expected Behavior**: Returns HTTP 200 with a new Access Token; the rotated Refresh Token is set only as an HTTP-only cookie.
      operationId: refreshAccessToken
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Optional explicit refresh token string if HTTP-only cookie is not present
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        "200":
          description: Tokens refreshed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Access Token refreshed
                data:
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset email
      description: >
        **Purpose**: Initiates password recovery process for a registered user.<br>
        **Authorization**: Public.<br>
        **Side Effects**: Generates temporary SHA-256 reset token with 20-minute expiration and dispatches email with reset link.<br>
        **Expected Behavior**: Always returns HTTP 200, regardless of whether the email belongs to an account.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        "200":
          description: Password reset email dispatched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: If an account exists, a reset email has been sent.
                data: null
        "400":
          $ref: '#/components/responses/ValidationError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /auth/reset-password/{resetToken}:
    post:
      tags:
        - Authentication
      summary: Reset password using reset token
      description: >
        **Purpose**: Sets a new account password using a valid password reset token.<br>
        **Authorization**: Public.<br>
        **Side Effects**: Replaces existing hashed password, invalidates reset token, and clears token expiration.<br>
        **Expected Behavior**: Returns HTTP 200 on successful password update. Throws HTTP 400 if token is invalid, expired, or passwords do not match.
      operationId: resetPassword
      parameters:
        - name: resetToken
          in: path
          required: true
          description: Temporary password reset token from email link
          schema:
            type: string
          example: 9f8e7d6c5b4a3f2e1d0c9b8a7f6e5d4c3b2a1f0e
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        "200":
          description: Password reset successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Password reset successfully
                data: null
        "400":
          $ref: '#/components/responses/ValidationError'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout current user
      description: >
        **Purpose**: Logs out the current user session.<br>
        **Authorization**: Required (`Bearer JWT`).<br>
        **Side Effects**: Removes `refreshToken` from user document in MongoDB and clears authentication cookies (`accessToken` and `refreshToken`).<br>
        **Expected Behavior**: Returns HTTP 200 confirming user logged out.
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful and cookies cleared.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: User logged out successfully
                data: null
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /auth/resend-email-verification:
    post:
      tags:
        - Authentication
      summary: Resend email verification link
      description: >
        **Purpose**: Re-generates and sends a new email verification link to the authenticated user's email address.<br>
        **Authorization**: Required (`Bearer JWT`).<br>
        **Side Effects**: Updates verification token and expiration in database and sends email.<br>
        **Expected Behavior**: Returns HTTP 200. Throws HTTP 400 if user email is already verified.
      operationId: resendVerificationEmail
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Email verification link dispatched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Verification email sent successfully
                data:
                  email: alex.rivera@taskflow.dev
        "400":
          description: User email is already verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 400
                success: false
                message: Email is already verified
                errors: []
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change account password
      description: >
        **Purpose**: Updates the password for an active authenticated user account.<br>
        **Authorization**: Required (`Bearer JWT`).<br>
        **Side Effects**: Verifies old password, updates hashed password in MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200. Throws HTTP 400 if old password is incorrect or new password validation fails.
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        "200":
          description: Password updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Password changed successfully
                data: null
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter valid JWT Access Token obtained from `/api/v1/auth/login`.
  schemas:
...
    RegisterUserRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 5
          maxLength: 40
          example: alex_rivera
          description: Unique account handle (5-40 characters, lowercase)
        email:
          type: string
          format: email
          example: alex.rivera@taskflow.dev
          description: Valid primary email address
        password:
          type: string
          format: password
          minLength: 8
          example: SecretPass123!
          description: User account password (minimum 8 characters)
    LoginUserRequest:
      type: object
      required:
        - password
      properties:
        username:
          type: string
          minLength: 5
          example: alex_rivera
          description: Registered username (lowercase)
        email:
          type: string
          format: email
          example: alex.rivera@taskflow.dev
          description: Registered email address
        password:
          type: string
          format: password
          example: SecretPass123!
          description: Account password
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: alex.rivera@taskflow.dev
          description: Registered account email address
    ResetPasswordRequest:
      type: object
      required:
        - newPassword
        - confirmNewPassword
      properties:
        newPassword:
          type: string
          format: password
          minLength: 8
          example: NewSecurePass123!
          description: New password (minimum 8 characters)
        confirmNewPassword:
          type: string
          format: password
          example: NewSecurePass123!
          description: Confirmation matching new password
    ChangePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
        - confirmNewPassword
      properties:
        oldPassword:
          type: string
          format: password
          example: SecretPass123!
          description: Current account password
        newPassword:
          type: string
          format: password
          minLength: 8
          example: NewSecurePass123!
          description: New password (minimum 8 characters)
        confirmNewPassword:
          type: string
          format: password
          example: NewSecurePass123!
          description: Confirmation matching new password
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - success
        - message
        - errors
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP response status code
        success:
          type: boolean
          example: false
          description: Success flag (false for errors)
        message:
          type: string
          example: Request failed due to invalid input validation.
          description: Error message summary
        errors:
          type: array
          description: Array of detailed validation error objects or strings
          items:
            type: object
          example:
            - field: email
              message: Invalid email format
        stack:
          type: string
          nullable: true
          example: "Error: Bad Request at..."
          description: Stack trace snippet (only populated in development environment)
    SuccessResponse:
      type: object
      required:
        - statusCode
        - success
        - message
        - data
      properties:
        statusCode:
          type: integer
          example: 200
          description: HTTP response status code
        success:
          type: boolean
          example: true
          description: Success flag (true for successful responses)
        message:
          type: string
          example: Operation completed successfully
          description: Success message summary
        data:
          nullable: true
          description: Response data payload. Its exact type is documented for every endpoint example and in FRONTEND_API_CONTRACT.md.
...
  responses:
    Unauthorized:
      description: Unauthorized - JWT Access Token is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            success: false
            message: Unauthorized request or invalid JWT token
            errors: []
...
    NotFound:
      description: Not Found - The requested resource (user, project, task, subtask, note) does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            success: false
            message: Resource not found
            errors: []
    Conflict:
      description: Conflict - Operation conflicts with existing state (e.g. duplicate username, email, or existing project member).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 409
            success: false
            message: User with this email or username already exists
            errors: []
    UnprocessableEntity:
      description: Unprocessable Entity - Request validation failed or parameters do not meet requirements.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 422
            success: false
            message: Validation failed for request fields
            errors:
              - field: confirmNewPassword
                message: Passwords do not match
    InternalServerError:
      description: Internal Server Error - An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 500
            success: false
            message: Internal Server Error
            errors: []
    ValidationError:
      description: Validation Error - Input parameters or payload failed express-validator rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 400
            success: false
            message: Validation Error
            errors:
              - field: email
                message: Invalid email format